feat: add switch org command
